// LogisticServerUI1/src/app/shared/models/entities/server-settings/enums/sscc-regulatory.enum.ts
export enum RegulatoryEnum {
    DGFT = "DGFT",
    HDMA = "HDMA",
    SFDA = "SFDA",
    EU = "EU",
    RUSSIAN = "RUSSIAN",
    CFDA = "CFDA",
    SOUTH_KORIA = "SOUTH KORIA",
    TURKEY = "TURKEY",
    ANVISA = "ANVISA",
    ARGENTINA = "ARGENTINA",
    EGYPT = "EGYPT",
    BAHRAIN = "BAHRAIN",
    DHA = "DHA",
    UAE = "UAE"
}